move gif_to_frames from misc to image converter, remove mist
