+++ core/movie_converter.py
@@ -8,7 +8,7 @@ from PIL import Image
 from tqdm import tqdm
 from utils.category import Category
 from core.doc_converter import office_to_frames
-from core.utils.misc import end_with_msg
+from core.utils.exit import end_with_msg
 from moviepy import (
     VideoFileClip,
     ImageClip,
